requirements.txt
aiogram==3.0.0b7
tortoise-orm==0.19.3
aerich==0.7.1
python-dotenv==1.0.0
pydantic==1.10.2
qrcode[pil]==7.4.2
cryptography==39.0.2
requests==2.28.2
python-dateutil==2.8.2
babel==2.12.1